feat: update kill command

+++ lc-hax/Scripts/Commands/KillCommand.cs
@@ -7,9 +7,19 @@ namespace Hax;
 public class KillCommand : ICommand {
     public void Execute(string[] args) {
         if (args.Length < 1) {
+            if (!Helpers.Extant(Helpers.LocalPlayer, out PlayerControllerB localPlayer)) {
+                Console.Print("SYSTEM", "Player not found!");
+                return;
+            }
+
+            localPlayer.DamagePlayerFromOtherClientServerRpc(int.MaxValue, Vector3.zero, -1);
+            return;
+        }
+
+        if (args[0] is "--all") {
             Helpers.Players
                    .ToList()
-                   .ForEach(player => player.DamagePlayerFromOtherClientServerRpc(1000, Vector3.zero, -1));
+                   .ForEach(player => player.DamagePlayerFromOtherClientServerRpc(int.MaxValue, Vector3.zero, -1));
 
             Console.Print("SYSTEM", "Attempting to kill all players!");
             return;
@@ -20,7 +30,7 @@ public class KillCommand : ICommand {
             return;
         }
 
-        targetPlayer.DamagePlayerFromOtherClientServerRpc(1000, Vector3.zero, -1);
+        targetPlayer.DamagePlayerFromOtherClientServerRpc(int.MaxValue, Vector3.zero, -1);
         Console.Print("SYSTEM", $"Attempting to kill {targetPlayer.playerUsername}!");
     }
 }
